+++ cmd/validator/wallet/create_test.go
@@ -1,19 +1,17 @@
-package accounts
+package wallet
 
 import (
 	"context"
 	"flag"
 	"io"
-	"os"
-	"path/filepath"
-	"strconv"
 	"testing"
 
 	"github.com/pkg/errors"
+	cmdacc "github.com/prysmaticlabs/prysm/v3/cmd/validator/accounts"
 	"github.com/prysmaticlabs/prysm/v3/cmd/validator/flags"
-	"github.com/prysmaticlabs/prysm/v3/config/params"
 	"github.com/prysmaticlabs/prysm/v3/testing/assert"
 	"github.com/prysmaticlabs/prysm/v3/testing/require"
+	"github.com/prysmaticlabs/prysm/v3/validator/accounts"
 	"github.com/prysmaticlabs/prysm/v3/validator/accounts/wallet"
 	"github.com/prysmaticlabs/prysm/v3/validator/keymanager"
 	"github.com/prysmaticlabs/prysm/v3/validator/keymanager/local"
@@ -23,87 +21,9 @@ import (
 	"github.com/urfave/cli/v2"
 )
 
-const (
-	passwordFileName = "password.txt"
-	password         = "OhWOWthisisatest42!$"
-)
-
 func init() {
 	logrus.SetLevel(logrus.DebugLevel)
 	logrus.SetOutput(io.Discard)
-}
-
-type testWalletConfig struct {
-	exitAll                 bool
-	skipDepositConfirm      bool
-	keymanagerKind          keymanager.Kind
-	numAccounts             int64
-	grpcHeaders             string
-	privateKeyFile          string
-	accountPasswordFile     string
-	walletPasswordFile      string
-	backupPasswordFile      string
-	backupPublicKeys        string
-	voluntaryExitPublicKeys string
-	deletePublicKeys        string
-	keysDir                 string
-	backupDir               string
-	passwordsDir            string
-	walletDir               string
-}
-
-func setupWalletCtx(
-	tb testing.TB,
-	cfg *testWalletConfig,
-) *cli.Context {
-	app := cli.App{}
-	set := flag.NewFlagSet("test", 0)
-	set.String(flags.WalletDirFlag.Name, cfg.walletDir, "")
-	set.String(flags.KeysDirFlag.Name, cfg.keysDir, "")
-	set.String(flags.KeymanagerKindFlag.Name, cfg.keymanagerKind.String(), "")
-	set.String(flags.DeletePublicKeysFlag.Name, cfg.deletePublicKeys, "")
-	set.String(flags.VoluntaryExitPublicKeysFlag.Name, cfg.voluntaryExitPublicKeys, "")
-	set.String(flags.BackupDirFlag.Name, cfg.backupDir, "")
-	set.String(flags.BackupPasswordFile.Name, cfg.backupPasswordFile, "")
-	set.String(flags.BackupPublicKeysFlag.Name, cfg.backupPublicKeys, "")
-	set.String(flags.WalletPasswordFileFlag.Name, cfg.walletPasswordFile, "")
-	set.String(flags.AccountPasswordFileFlag.Name, cfg.accountPasswordFile, "")
-	set.Int64(flags.NumAccountsFlag.Name, cfg.numAccounts, "")
-	set.Bool(flags.SkipDepositConfirmationFlag.Name, cfg.skipDepositConfirm, "")
-	set.Bool(flags.SkipMnemonic25thWordCheckFlag.Name, true, "")
-	set.Bool(flags.ExitAllFlag.Name, cfg.exitAll, "")
-	set.String(flags.GrpcHeadersFlag.Name, cfg.grpcHeaders, "")
-
-	if cfg.privateKeyFile != "" {
-		set.String(flags.ImportPrivateKeyFileFlag.Name, cfg.privateKeyFile, "")
-		assert.NoError(tb, set.Set(flags.ImportPrivateKeyFileFlag.Name, cfg.privateKeyFile))
-	}
-	assert.NoError(tb, set.Set(flags.WalletDirFlag.Name, cfg.walletDir))
-	assert.NoError(tb, set.Set(flags.SkipMnemonic25thWordCheckFlag.Name, "true"))
-	assert.NoError(tb, set.Set(flags.KeysDirFlag.Name, cfg.keysDir))
-	assert.NoError(tb, set.Set(flags.KeymanagerKindFlag.Name, cfg.keymanagerKind.String()))
-	assert.NoError(tb, set.Set(flags.DeletePublicKeysFlag.Name, cfg.deletePublicKeys))
-	assert.NoError(tb, set.Set(flags.VoluntaryExitPublicKeysFlag.Name, cfg.voluntaryExitPublicKeys))
-	assert.NoError(tb, set.Set(flags.BackupDirFlag.Name, cfg.backupDir))
-	assert.NoError(tb, set.Set(flags.BackupPublicKeysFlag.Name, cfg.backupPublicKeys))
-	assert.NoError(tb, set.Set(flags.BackupPasswordFile.Name, cfg.backupPasswordFile))
-	assert.NoError(tb, set.Set(flags.WalletPasswordFileFlag.Name, cfg.walletPasswordFile))
-	assert.NoError(tb, set.Set(flags.AccountPasswordFileFlag.Name, cfg.accountPasswordFile))
-	assert.NoError(tb, set.Set(flags.NumAccountsFlag.Name, strconv.Itoa(int(cfg.numAccounts))))
-	assert.NoError(tb, set.Set(flags.SkipDepositConfirmationFlag.Name, strconv.FormatBool(cfg.skipDepositConfirm)))
-	assert.NoError(tb, set.Set(flags.ExitAllFlag.Name, strconv.FormatBool(cfg.exitAll)))
-	assert.NoError(tb, set.Set(flags.GrpcHeadersFlag.Name, cfg.grpcHeaders))
-	return cli.NewContext(&app, set, nil)
-}
-
-func setupWalletAndPasswordsDir(t testing.TB) (string, string, string) {
-	walletDir := filepath.Join(t.TempDir(), "wallet")
-	passwordsDir := filepath.Join(t.TempDir(), "passwords")
-	passwordFileDir := filepath.Join(t.TempDir(), "passwordFile")
-	require.NoError(t, os.MkdirAll(passwordFileDir, params.BeaconIoConfig().ReadWriteExecutePermissions))
-	passwordFilePath := filepath.Join(passwordFileDir, passwordFileName)
-	require.NoError(t, os.WriteFile(passwordFilePath, []byte(password), os.ModePerm))
-	return walletDir, passwordsDir, passwordFilePath
 }
 
 func TestCreateOrOpenWallet(t *testing.T) {
@@ -116,19 +36,19 @@ func TestCreateOrOpenWallet(t *testing.T) {
 		walletPasswordFile: walletPasswordFile,
 	})
 	createLocalWallet := func(cliCtx *cli.Context) (*wallet.Wallet, error) {
-		cfg, err := ExtractWalletCreationConfigFromCli(cliCtx, keymanager.Local)
+		cfg, err := cmdacc.ExtractWalletDirPassword(cliCtx)
 		if err != nil {
 			return nil, err
 		}
 		w := wallet.New(&wallet.Config{
-			KeymanagerKind: cfg.WalletCfg.KeymanagerKind,
-			WalletDir:      cfg.WalletCfg.WalletDir,
-			WalletPassword: cfg.WalletCfg.WalletPassword,
+			KeymanagerKind: keymanager.Local,
+			WalletDir:      cfg.Dir,
+			WalletPassword: cfg.Password,
 		})
-		if err = CreateLocalKeymanagerWallet(cliCtx.Context, w); err != nil {
+		if err = accounts.CreateLocalKeymanagerWallet(cliCtx.Context, w); err != nil {
 			return nil, errors.Wrap(err, "could not create keymanager")
 		}
-		log.WithField("wallet-path", cfg.WalletCfg.WalletDir).Info(
+		log.WithField("wallet-path", cfg.Dir).Info(
 			"Successfully created new wallet",
 		)
 		return w, nil
