+++ validator/accounts/BUILD.bazel
@@ -45,7 +45,6 @@ go_library(
         "//validator/keymanager/derived:go_default_library",
         "//validator/keymanager/local:go_default_library",
         "//validator/keymanager/remote:go_default_library",
-        "//validator/keymanager/remote-web3signer:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
         "@com_github_google_uuid//:go_default_library",
         "@com_github_logrusorgru_aurora//:go_default_library",
@@ -65,7 +64,6 @@ go_test(
         "accounts_exit_test.go",
         "accounts_import_test.go",
         "accounts_list_test.go",
-        "wallet_create_test.go",
         "wallet_recover_fuzz_test.go",
         "wallet_recover_test.go",
     ],
@@ -96,7 +94,6 @@ go_test(
         "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_google_uuid//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_sirupsen_logrus//hooks/test:go_default_library",
         "@com_github_urfave_cli_v2//:go_default_library",
         "@com_github_wealdtech_go_eth2_wallet_encryptor_keystorev4//:go_default_library",
