+++ validator/node/BUILD.bazel
@@ -15,7 +15,6 @@ go_test(
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
         "//validator/accounts:go_default_library",
-        "//validator/accounts/wallet:go_default_library",
         "//validator/keymanager:go_default_library",
         "//validator/keymanager/remote-web3signer:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
